Separate out pystol.cpp into a couple files
